requirements.txt
Flask==2.3.6
gunicorn==21.2.0
pandas==2.1.1
soundfile==0.12.1
scipy==1.12.1
openpyxl==3.1.3
gspread==5.8.2
google-api-python-client==2.93.0
google-auth==2.24.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0